mm: migrate: close race between migration completion and mprotect
A migration entry is marked as write if pte_write was true at the time the
entry was created.  The VMA protections are not double checked when
migration entries are being removed as mprotect marks
write-migration-entries as read.  It means that potentially we take a
spurious fault to mark PTEs write again but it's straight-forward. 
However, there is a race between write migrations being marked read and
migrations finishing.  This potentially allows a PTE to be write that
should have been read.  Close this race by double checking the VMA
permissions using maybe_mkwrite when migration completes.

[torvalds@linux-foundation.org: use maybe_mkwrite]
Signed-off-by: Mel Gorman <mgorman@suse.de>
Acked-by: Rik van Riel <riel@redhat.com>
Cc: Dave Jones <davej@redhat.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Michel Lespinasse <walken@google.com>
Cc: Kirill A Shutemov <kirill.shutemov@linux.intel.com>
Cc: Sasha Levin <sasha.levin@oracle.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>