mm: memcontrol: do not filter reclaimable nodes in NUMA round-robin
The round-robin node reclaim currently tries to include only nodes that
have memory of the memcg in question, which is quite elaborate.

Just use plain round-robin over the nodes that are allowed by the task's
cpuset, which are the most likely to contain that memcg's memory.  But
even if zones without memcg memory are encountered, direct reclaim will
skip over them without too much hassle.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Acked-by: Vladimir Davydov <vdavydov@parallels.com>
Cc: Michal Hocko <mhocko@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>