mm, compaction: prevent infinite loop in compact_zone
Several people have reported occasionally seeing processes stuck in
compact_zone(), even triggering soft lockups, in 3.18-rc2+.

Testing a revert of e14c720efdd7 ("mm, compaction: remember position
within pageblock in free pages scanner") fixed the issue, although the
stuck processes do not appear to involve the free scanner.  Finally, by
code inspection, the bug was found in isolate_migratepages() which uses a
slightly different condition to detect if the migration and free scanners
have met, than compact_finished().  That has not been a problem until
commit e14c720efdd7 allowed the free scanner position between individual
invocations to be in the middle of a pageblock.

In a relatively rare case, the migration scanner position can end up at
the beginning of a pageblock, with the free scanner position in the middle
of the same pageblock.  If it's the migration scanner's turn,
isolate_migratepages() exits immediately (without updating the position),
while compact_finished() decides to continue compaction, resulting in a
potentially infinite loop.  The system can recover only if another process
creates enough high-order pages to make the watermark checks in
compact_finished() pass.

This patch fixes the immediate problem by bumping the migration scanner's
position to meet the free scanner in isolate_migratepages(), when both are
within the same pageblock.  This causes compact_finished() to terminate
properly.  A more robust check in compact_finished() is planned as a
cleanup for better future maintainability.

Fixes: e14c720efdd73 ("mm, compaction: remember position within pageblock in free pages scanner)
Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Reported-by: P. Christeas <xrg@linux.gr>
Tested-by: P. Christeas <xrg@linux.gr>
Link: http://marc.info/?l=linux-mm&m=141508604232522&w=2
Reported-by: Norbert Preining <preining@logic.at>
Tested-by: Norbert Preining <preining@logic.at>
Link: https://lkml.org/lkml/2014/11/4/904
Reported-by: Pavel Machek <pavel@ucw.cz>
Link: https://lkml.org/lkml/2014/11/7/164
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>