mm-compaction-pass-classzone_idx-and-alloc_flags-to-watermark-checking-fix
This patch-fix changes zone watermark checking in compaction_suitable() as
we discussed with Joonsoo Kim.  First, it moves up the watermark check for
answering the question "does the zone need compaction at all?".  Before
this change, the check is preceded by a check that answers the question
"does the zone have enough free pages to succeed compaction".  So it might
happen that there is already a high-order page available, but not enough
pages for performing the compaction (which assumes extra pages for the
migrations).  Before the patch, compaction_suitable() would return
COMPACT_SKIPPED which means "compaction cannot succeed, reclaim more",
after this change it returns COMPACT_PARTIAL which means "compaction not
needed, try allocating".

Second, the check for answering "can the compaction succeed?" now also
includes classzone_idx and alloc_flags parameters.  This prevents starting
compaction that would not lead to successful allocation due to not having
enough free pages.  The addition of extra pages for migration is left in
the check.  Although these are temporary allocations and thus should not
be affected by alloc_flags and classzone_idx, it only matters when we are
close to the watermark, where it does not hurt to be a bit pessimistic.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Michal Nazarewicz <mina86@mina86.com>
Cc: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>