hwmon: qpnp-adc: Remove function qpnp_vadc_check_result()
The function qpnp_vadc_check_result() checks the code against
ideal min/max for 0V and 1.8V and truncates the result within
these values. However if there is a positive offset the ADC
value can be greater than 0xA800 when applying a voltage of
1.8V. Similarly if the offset is negative the value can be
less than 0x6000 when applying 0V. Having this check can
result in incorrect ratiometric calibration if the value
read for 0V and 1.8V is outside the ideal ADC code and
the value is truncated by this function call.

Change-Id: If0c89ac4bffdea5438e256233860b2aca80e826b
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
Signed-off-by: Kranthikumar Kurapati <kkurap@codeaurora.org>